requirements.txt
fastapi==0.111.0
pydantic_settings==2.3.4
Requests==2.32.3
streamlit==1.36.0
uvicorn==0.30.1
python-dotenv~=1.0.0
pysqlite3
mistralai
